requirements.txt
streamlit
yfinance
pandas
matplotlib
fpdf2
vaderSentiment
praw
